refactor(session): single teardownTabs core for kill + release-PTY (#1195 Phase 2a)
Session teardown was open-coded three times — LocalBackend.Kill,
LocalBackend.CloseAttachOnly, and Instance.ArchiveTeardown — each with its
own local tabSession struct and the same "snapshot tmux under i.mu → close
outside the lock → re-lock, clear refs" skeleton, diverging in four silent
ways (worktree handling, the #802 pane-exit-before-worktree ordering, the
started fence, and the ref-clear identity guard). That divergence is the bug
surface behind the #1196 archive-orphan class.

Phase 2a introduces the shared core teardownTabs(mode) with the KILL and
RELEASE-PTY modes and migrates the two backend methods onto it, collapsing
two of the three copied loops. Behavior-preserving:

- kill mode reproduces LocalBackend.Kill exactly — kill-session + wait for
  each pane to exit (#802), then delete the worktree, then clear refs; the
  `kill "<title>":` warning prefix is kept verbatim (backend_test asserts it).
- release-PTY mode reproduces LocalBackend.CloseAttachOnly — drop every tab's
  attach PTY without kill-session, leave the worktree untouched, return the
  joined per-tab errors (#867/#1065).
- ref-clearing is identity-guarded for both modes (a strict refinement of
  Kill's old unconditional clear: identical result absent a concurrent Start
  swap, correct in the race Kill didn't handle).

Phase 2b adds the ARCHIVE mode (folding MoveWorktree into the core right after
the pane-exit wait, closing the #802 duplication) and migrates
ArchiveTeardown. Refs #1195.
